ingest/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
neo4j==5.14.0
pinecone-client==2.2.4
sentence-transformers==2.2.2
tqdm==4.66.1
